examples/rp: 'Greedy' sample to demonstrated deadlock with multiple waiters
This example demonstrates that using any primitive that uses
WakerRegistration will cause a tight-poll loop if more than one task
waits for it.

The comments suggest this is known, but also that "this will still
work".  This test demonstrates that if the task that causes the actual
wake is at a lower priority, it will never get a chance to run.